requirements.txt
requests==2.31.0
python-dotenv==1.0.0
Pillow==11.3.0
matplotlib==3.7.2 